Fix port conflict when running multiple unit tests in parallel
