+++ src/main/java/io/specto/hoverfly/junit/api/HoverflyClientFactory.java
@@ -1,0 +1,12 @@
+package io.specto.hoverfly.junit.api;
+
+import io.specto.hoverfly.junit.core.config.HoverflyConfiguration;
+
+/**
+ * Factory for creating {@link HoverflyClient}s
+ */
+public interface HoverflyClientFactory {
+
+    HoverflyClient createHoverflyClient(HoverflyConfiguration hoverflyConfig);
+
+}
