+++ src/test/java/io/specto/hoverfly/junit/core/HoverflyTest.java
@@ -10,8 +10,10 @@ import com.google.common.collect.Sets;
 import com.google.common.io.Resources;
 import io.specto.hoverfly.junit.api.HoverflyClient;
 import io.specto.hoverfly.junit.api.HoverflyClientException;
+import io.specto.hoverfly.junit.api.HoverflyClientFactory;
 import io.specto.hoverfly.junit.api.model.ModeArguments;
 import io.specto.hoverfly.junit.api.view.HoverflyInfoView;
+import io.specto.hoverfly.junit.core.config.HoverflyConfiguration;
 import io.specto.hoverfly.junit.core.config.LocalHoverflyConfig;
 import io.specto.hoverfly.junit.core.config.LogLevel;
 import io.specto.hoverfly.junit.core.model.DelaySettings;
@@ -228,7 +230,6 @@ public class HoverflyTest {
     }
 
 
-
     @Test
     public void shouldThrowExceptionWhenExportSimulationWithoutPath() {
 
@@ -397,9 +398,9 @@ public class HoverflyTest {
     public void shouldResetSimulationJournalAndStateWhenCallingReset() {
 
         hoverfly = new Hoverfly(SIMULATE);
-
         HoverflyClient hoverflyClient = createMockHoverflyClient(hoverfly);
         when(hoverflyClient.getHealth()).thenReturn(true);
+        hoverfly.start();
 
         hoverfly.reset();
 
@@ -409,7 +410,7 @@ public class HoverflyTest {
     }
 
     @Test
-    public void shouldCopySslCertAndKeyToTempFolderIfPresent () {
+    public void shouldCopySslCertAndKeyToTempFolderIfPresent() {
         // Given
         hoverfly = new Hoverfly(localConfigs()
                 .sslCertificatePath("ssl/ca.crt")
@@ -426,13 +427,13 @@ public class HoverflyTest {
     }
 
     @Test
-    public void shouldCopyMiddlewareScriptToTempFolderIfLocalMiddlewareEnabled () {
+    public void shouldCopyMiddlewareScriptToTempFolderIfLocalMiddlewareEnabled() {
         String rawFilename = "middleware.py";
         String path = "middleware" + File.separator + rawFilename;
 
         // Given
         hoverfly = new Hoverfly(localConfigs()
-           .localMiddleware("python", path), SIMULATE);
+                .localMiddleware("python", path), SIMULATE);
         TempFileManager tempFileManager = spy(TempFileManager.class);
         Whitebox.setInternalState(hoverfly, "tempFileManager", tempFileManager);
 
@@ -460,10 +461,13 @@ public class HoverflyTest {
     }
 
     @Test
-    public void shouldValidateHoverflyConfigBeforeStart() {
+    public void shouldFindUnusedPortWhenStarted() {
 
         hoverfly = new Hoverfly(SIMULATE);
+        assertThat(hoverfly.getHoverflyConfig().getProxyPort()).isZero();
+        assertThat(hoverfly.getHoverflyConfig().getAdminPort()).isZero();
 
+        hoverfly.start();
         assertThat(hoverfly.getHoverflyConfig().getProxyPort()).isNotZero();
         assertThat(hoverfly.getHoverflyConfig().getAdminPort()).isNotZero();
     }
@@ -566,18 +570,18 @@ public class HoverflyTest {
     @Test
     public void shouldResetModeWithModeArguments() {
         hoverfly = new Hoverfly(localConfigs().captureAllHeaders().enableStatefulCapture(), CAPTURE);
-
         HoverflyClient hoverflyClient = createMockHoverflyClient(hoverfly);
+        when(hoverflyClient.getHealth()).thenReturn(true);
+        hoverfly.start();
 
         hoverfly.resetMode(CAPTURE);
 
         final ArgumentCaptor<ModeArguments> arguments = ArgumentCaptor.forClass(ModeArguments.class);
-        verify(hoverflyClient).setMode(eq(HoverflyMode.CAPTURE), arguments.capture());
+        verify(hoverflyClient, times(2)).setMode(eq(HoverflyMode.CAPTURE), arguments.capture());
 
         assertThat(arguments.getValue().isStateful()).isTrue();
         assertThat(arguments.getValue().getHeadersWhitelist()).containsOnly("*");
     }
-
 
 
     @Test
@@ -595,7 +599,9 @@ public class HoverflyTest {
 
         hoverfly = new Hoverfly(SIMULATE);
         HoverflyClient hoverflyClient = createMockHoverflyClient(hoverfly);
+        when(hoverflyClient.getHealth()).thenReturn(true);
         doThrow(HoverflyClientException.class).when(hoverflyClient).deleteJournal();
+        hoverfly.start();
 
         hoverfly.resetJournal();
 
@@ -620,7 +626,9 @@ public class HoverflyTest {
         // given
         hoverfly = new Hoverfly(DIFF);
         HoverflyClient hoverflyClient = createMockHoverflyClient(hoverfly);
+        when(hoverflyClient.getHealth()).thenReturn(true);
         doThrow(HoverflyClientException.class).when(hoverflyClient).cleanDiffs();
+        hoverfly.start();
 
         // when
         hoverfly.resetDiffs();
@@ -711,7 +719,9 @@ public class HoverflyTest {
 
     private HoverflyClient createMockHoverflyClient(Hoverfly hoverfly) {
         HoverflyClient hoverflyClient = mock(HoverflyClient.class);
-        Whitebox.setInternalState(hoverfly, "hoverflyClient", hoverflyClient);
+        HoverflyClientFactory clientFactory = mock(HoverflyClientFactory.class);
+        when(clientFactory.createHoverflyClient(any(HoverflyConfiguration.class))).thenReturn(hoverflyClient);
+        Whitebox.setInternalState(hoverfly, "hoverflyClientFactory", clientFactory);
         return hoverflyClient;
     }
